% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintTweetsComparation.R
\name{paintTweetsComparation}
\alias{paintTweetsComparation}
\title{Tweets Comparation}
\usage{
paintTweetsComparation(file, file2, file3)
}
\arguments{
\item{file}{A data frame with the attack information.}

\item{file2}{A data frame with the attack 2 information.}

\item{file3}{A data frame with the attack 3 information.}
}
\description{
Tweets Comparation
}
