% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintWorldMap.R
\name{paintWorldMap}
\alias{paintWorldMap}
\title{Geocode World Map}
\usage{
paintWorldMap(file)
}
\arguments{
\item{file}{A data frame with the attack information.}
}
\description{
Paint the tweets geogrhapic codes information on a global map.
}
