% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordCloudTweets.R
\name{wordCloudTweets}
\alias{wordCloudTweets}
\title{word Cloud Tweets}
\usage{
wordCloudTweets(file)
}
\arguments{
\item{file}{A data frame with the attack information.}
}
\description{
word Cloud Tweets
}
